converted the normalized score to a percentage (whole number)
- the conversion is done at the model level
- on the interface, we appended '%' for more clarity
- in the report we added 'Percentage Score for' at the front of
each criteria to let the user know what the numbers mean
	- if we just append '%' to the end of the numbers, Excel
	and text editors seem to append decimal places to them.

+++ acj/dataformat.py
@@ -258,5 +258,5 @@ def getScores():
 		'rounds': fields.Integer,
 		'wins': fields.Integer,
 		'score': fields.Float,
-		'normalized_score': fields.Float
+		'normalized_score': fields.Integer
 	}
